requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.4.0
attrs==23.2.0
az==0.1.0.dev1
azure-cognitiveservices-speech==1.38.0
azure-core==1.30.2
azure-identity==1.17.1
azure-monitor-opentelemetry-exporter==1.0.0b27
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
docutils==0.21.2
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
filetype==1.2.0
fixedint==0.1.6
Flask==3.0.3
Flask-Cors==4.0.1
flask-restx==1.3.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
more-itertools==10.3.0
msal==1.29.0
msal-extensions==1.2.0
msrest==0.7.1
numpy==2.0.0
oauthlib==3.2.2
openai==1.35.13
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
portalocker==2.10.0
promptflow==1.13.0
promptflow-core==1.13.0
promptflow-devkit==1.13.0
promptflow-tracing==1.13.0
protobuf==4.25.3
psutil==5.9.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydash==7.0.7
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
strictyaml==1.7.3
tabulate==0.9.0
tiktoken==0.7.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
waitress==2.1.2
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2